Move db related files in server/db
