+++ src/app/api/trpc/[trpc]/route.ts
@@ -1,4 +1,4 @@
-import { env } from "@/env.mjs";
+import { env } from "@/env";
 import { createContext } from "@/server/api/context";
 import { appRouter } from "@/server/api/root";
 import { fetchRequestHandler } from "@trpc/server/adapters/fetch";
