+++ tsconfig.json
@@ -39,5 +39,5 @@
 		"**/*.js",
 		".next/types/**/*.ts"
 	],
-	"exclude": ["node_modules"]
+	"exclude": ["node_modules", "public"]
 }
